early-init.el: tweak default height
When I was testing this before, the lower value was necessary otherwise
the frame would not fit into the screen and it would change into an
unpredictable size.  I guess some change upstream has since fixed this?
Now setting the height to be equal to my monitor's works even if there
is a system panel and a titlebar for the frame.