prot-spell: use unique name for Ispell corrections
This is an attempt to have a sufficiently obscure name for what was the
"*Choices*" buffer so as to target it by name alone in the setup of
display-buffer-alist.

I thought the previous arrangement would work, but Fourchaux reports
that it gives false positives.  Hopefully the issue is now fixed:
<https://gitlab.com/protesilaos/dotfiles/-/issues/55>.