prot-vc: BREAKING remove the minor mode
I am repurposing this file.  I want it to just provide commands that I
need to use.

I do not need the fancier setup anymore, because diff-mode can now
interpret C-x v v to commit the contents of the diff.  When that
happens, the diff buffer stays in place and the log-edit buffers pops
up in a new window.  So the layout I once had is derived that way
organically.

A diff buffer can always be viewed on demand from inside the log-edit
buffer with C-c C-d (same with Magit's git commit, by the way).