sway: tweak keyboard to bind ESC to CAPS
This is because I am currently using evil-mode (Vim keys) in Emacs.  I
also switched to Sway for the time being.  I was using GNOME before to
get a feel of how Wayland behaves: in short, it is great.  Sway seems
to be the best tiling Wayland compositor.  I noticed that "river" is
another option, but I will have to check that out at a later date.