prot-emacs-modus-themes: REWRITE for dev version
Use the package macro defined in 2320416b.  Use the theme functions of
08f4ccbb.

Version 4 of the modus-themes is in active development.  There is no
release of it yet and the code is in its own 'version-4' branch.

I need this rewrite to better test how package.el handles the new
changes and to make sure the package builds correctly.