bspwmrc: REMOVE emacs daemon
It is not reliable.

I was running a shell buffer and tried to evaluate some Elisp that was
erroneous.  Instead of throwing an error in the buffer, the whole
session crashed.  THIS HAS NEVER HAPPENED BEFORE.  I thought the error
was particularly egregious, but no, I experienced another crash with
something more innocuous.

I was running the daemon with the Lucid toolkit, as I read (in the 'man
emacs'?) that X11 can experience crashes...

This was horrible and I cannot recommend the daemon to anyone.