prot-emacs-completion-vertico: add back marginalia
It should be part of the Vertico bundle, not prot-emacs-completion-common.el.
The reason is that the *Completions* buffer is generated at the outset
with all the annotations included.  Extra annotations for stuff like
M-x and C-h o make the MCT experience very sluggish.  This is not the
fault of 'marginalia-mode'!  It simply is an inherent limitation of
the *Completions*.  As such, I am defining some annotations for a few
cases where I do not expect any noticeable slowdown.

I will probably end up modifying the 'marginalia-annotator-registry',
but I need to test this first.