prot-modeline: EXPERIMENT with built-in right align
I tried this before and it was not working as expected. Now it does on
the laptop, though I am not displaying the notmuch-indicator here.
Maybe this is what throws me off each time. I will check on the
desktop computer where I run notmuch and will have a better idea then.
Maybe I need to change things in notmuch-indicator.el. We will see.