emacs: stop reversing the vertico layout
It works fine most of the time, except when the text is too large and
the contents of the minibuffer consist of candidates with long lines.
There may be other issues as well, but this is the obvious one that I
would encounter frequently while doing 'consult-outline' on, say, my
prot-emacs.org file while using any 'fontaine' preset with a default
font size greater than 'medium'.